Remove unused vars, explicit integer casts, update screen sim paths

+++ firmware/src/pages/module_view.hh
@@ -113,7 +113,7 @@ struct ModuleViewPage : PageBase {
 
 		for (const auto el : moduleinfo.Switches) {
 			draw_switch(el, patch);
-			module_controls.push_back({ModuleParam::Type::Switch, el.id + moduleinfo.Knobs.size()});
+			module_controls.push_back({ModuleParam::Type::Switch, el.id + (uint32_t)moduleinfo.Knobs.size()});
 		}
 
 		// remove final \n
